make it better

+++ database/migrations/2019_10_14_103138_add_kenshinsys_column_to_hospitals.php
@@ -14,7 +14,7 @@ class AddKenshinsysColumnToHospitals extends Migration
     public function up()
     {
         Schema::table('hospitals', function (Blueprint $table) {
-            $table->integer('kenshin_sys_hospital_id')->nullable()->after('medical_examination_system_id');
+            $table->bigInteger('kenshin_sys_hospital_id')->nullable()->after('medical_examination_system_id');
         });
     }
